Warnings--
git-svn-id: https://rimeime.googlecode.com/svn/trunk/librime@348 b90aed45-4297-9963-ff9d-98b604ac09a9

+++ include/rime/candidate.h
@@ -19,15 +19,15 @@ class Candidate {
  public:
   Candidate() : type_(), start_(0), end_(0) {}
   Candidate(const std::string type,
-            int start,
-            int end) : type_(type), start_(start), end_(end) {}
+            size_t start,
+            size_t end) : type_(type), start_(start), end_(end) {}
   virtual ~Candidate() {}
 
   // recognized by translators in learning phase
   const std::string& type() const { return type_; }
   // [start, end) mark a range in the input that the candidate correspond to
-  int start() const { return start_; }
-  int end() const { return end_; }
+  size_t start() const { return start_; }
+  size_t end() const { return end_; }
   
   // candidate text to commit
   virtual const std::string& text() const = 0;
@@ -37,21 +37,21 @@ class Candidate {
   virtual const std::string preedit() const { return std::string(); }
 
   void set_type(const std::string &type) { type_ = type; }
-  void set_start(int start) { start_ = start; }
-  void set_end(int end) { end_ = end; }
+  void set_start(size_t start) { start_ = start; }
+  void set_end(size_t end) { end_ = end; }
 
  private:
   std::string type_;
-  int start_;
-  int end_;
+  size_t start_;
+  size_t end_;
 };
 
 class SimpleCandidate : public Candidate {
  public:
   SimpleCandidate() : Candidate() {}
   SimpleCandidate(const std::string type,
-                  int start,
-                  int end,
+                  size_t start,
+                  size_t end,
                   const std::string &text,
                   const std::string &comment = std::string(),
                   const std::string &preedit = std::string())
